requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.8
django-js-routes==0.2.0
django-vite==3.0.1
idna==3.6
inertia-django==0.5.3
marshmallow==3.20.1
packaging==23.2
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
